Eliminate float precision errors

+++ collect_app/src/main/java/org/odk/collect/android/widgets/range/RangeDecimalWidget.kt
@@ -65,7 +65,7 @@ class RangeDecimalWidget(
 
     override fun getAnswer(): IAnswerData? {
         return rangeSliderState.realValue?.let {
-            DecimalData(it.toDouble())
+            DecimalData(it)
         }
     }
